requirements.txt
# Main requirements for SAE analysis and sample generation
goodfire>=0.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.28.0

# Hugging Face for model downloads and local inference
huggingface-hub>=0.20.0
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.25.0

# For local model inference
sentencepiece>=0.1.99
protobuf>=4.21.0
safetensors>=0.3.0

# Note: vllm removed due to compilation issues on some systems
# Using transformers directly for local inference instead
